+++ src/platform/api/__init__.py
@@ -1,0 +1,5 @@
+"""API route groupings for platform entrypoints."""
+
+from .anthropic import router as anthropic_router
+
+__all__ = ["anthropic_router"]
