+++ src/platform/app.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 import json
 import logging
 from pathlib import Path
-import math
 from typing import Any, AsyncGenerator
 
 import httpx
@@ -17,6 +16,12 @@ from pydantic import AnyHttpUrl, BaseModel, ConfigDict, Field
 from urllib.parse import urlsplit, urlunsplit
 
 import src.transllm  # noqa: F401 - ensures adapters register with ProviderRegistry
+from src.platform.token_utils import (
+    _extract_content_text,
+    _safe_json,
+    _estimate_tokens,
+    _estimate_request_tokens,
+)
 from src.transllm.converters.response_converter import ResponseConverter
 from src.transllm.converters.stream_converter import StreamConverter
 from src.transllm.core.exceptions import ConversionError, UnsupportedProviderError
@@ -69,96 +74,6 @@ def _truncate_text(text: str, limit: int = 2000) -> str:
     if len(text) <= limit:
         return text
     return text[: limit - 3] + "..."
-
-
-def _safe_json(obj: Any) -> str:
-    try:
-        return json.dumps(obj, ensure_ascii=False)
-    except Exception:
-        return str(obj)
-
-
-def _extract_content_text(content: Any) -> str:
-    """Extract text-ish content from a unified content field."""
-    parts: list[str] = []
-    if isinstance(content, str):
-        parts.append(content)
-    elif isinstance(content, list):
-        for block in content:
-            if isinstance(block, ContentBlock):
-                if block.text:
-                    parts.append(block.text)
-                if block.reasoning and getattr(block.reasoning, "content", None):
-                    parts.append(str(block.reasoning.content))
-                if block.thinking and getattr(block.thinking, "content", None):
-                    parts.append(str(block.thinking.content))
-                if block.redacted_thinking and getattr(
-                    block.redacted_thinking, "content", None
-                ):
-                    parts.append(str(block.redacted_thinking.content))
-                if block.tool_result and getattr(block.tool_result, "result", None):
-                    parts.append(_safe_json(block.tool_result.result))
-                if block.image_url and getattr(block.image_url, "url", None):
-                    parts.append(block.image_url.url)
-            else:
-                # If a plain dict sneaks in, grab common fields
-                text_val = ""
-                if isinstance(block, dict):
-                    text_val = block.get("text") or block.get("reasoning") or ""
-                if text_val:
-                    parts.append(str(text_val))
-    return " ".join(parts)
-
-
-def _estimate_tokens(text: str, model: str | None = None) -> int:
-    """Best-effort token estimation with optional tiktoken, else heuristic."""
-    try:
-        import tiktoken  # type: ignore
-    except ImportError:
-        tiktoken = None  # type: ignore
-
-    if tiktoken:
-        try:
-            if model:
-                enc = tiktoken.encoding_for_model(model)
-            else:
-                enc = tiktoken.get_encoding("cl100k_base")
-            return len(enc.encode(text))
-        except Exception:
-            pass
-
-    # Heuristic fallback: ~4 chars per token, at least 1
-    return max(1, math.ceil(len(text) / 4))
-
-
-def _estimate_request_tokens(unified_request: Any) -> int:
-    """Estimate input tokens from a unified CoreRequest."""
-    model_name = getattr(unified_request, "model", None)
-    total_text: list[str] = []
-
-    sys_instr = getattr(unified_request, "system_instruction", None)
-    if sys_instr:
-        total_text.append(str(sys_instr))
-
-    for msg in getattr(unified_request, "messages", []) or []:
-        # role
-        role_val = (
-            msg.role.value if hasattr(msg.role, "value") else getattr(msg, "role", "")
-        )
-        if role_val:
-            total_text.append(str(role_val))
-
-        total_text.append(_extract_content_text(getattr(msg, "content", "")))
-
-        if getattr(msg, "tool_calls", None):
-            for tc in msg.tool_calls or []:
-                name = getattr(tc, "name", "")
-                args = getattr(tc, "arguments", {})
-                total_text.append(str(name))
-                total_text.append(_safe_json(args))
-
-    combined = " ".join(part for part in total_text if part)
-    return _estimate_tokens(combined, model=model_name)
 
 
 def _augment_message_start(
@@ -927,9 +842,18 @@ async def stream(
         "Connection": "keep-alive",
         "X-Accel-Buffering": "no",
     }
+    #
+    # return StreamingResponse(
+    #     event_publisher(),
+    #     media_type="text/event-stream",
+    #     headers=sse_headers,
+    # )
 
-    return StreamingResponse(
-        event_publisher(),
-        media_type="text/event-stream",
-        headers=sse_headers,
-    )
+    if client_provider is Provider.anthropic:
+        return StreamingResponse(
+            event_publisher(),
+            media_type="text/event-stream",
+            headers=sse_headers,
+        )
+    else:
+        pass
