+++ src/transllm/adapters/anthropic/__init__.py
@@ -1,6 +1,5 @@
 """Anthropic adapter and helpers for format conversion"""
 
 from .adapter import AnthropicAdapter
-from .client import AnthropicMessagesClient
 
-__all__ = ["AnthropicAdapter", "AnthropicMessagesClient"]
+__all__ = ["AnthropicAdapter", ]
